fix: refactor resources and modules
