+++ src/lib/pages/account-details/components/modules/index.ts
@@ -1,0 +1,1 @@
+export * from "./ModuleLists";
