+++ src/lib/pages/account-details/components/resources/index.ts
@@ -1,2 +1,2 @@
-export * from "./ResourceLists";
+export * from "./ResourceOverview";
 export * from "./ResourceSection";
